🛡️ B5ERD0 tests: split guard implementation suite
