+++ .agentplane/tasks/202605010645-B5ERD0/pr/github-title.txt
@@ -1,0 +1,1 @@
+AP-09: Split guard implementation tests (B5ERD0)
