+++ .agentplane/tasks/202605010645-B5ERD0/pr/review.md
@@ -1,0 +1,57 @@
+# PR Review
+
+Created: 2026-05-01T09:29:51.147Z
+Branch: task/202605010645-B5ERD0/guard-test-split
+
+## Summary
+
+AP-09: Split guard implementation tests
+
+Split guard commands.unit.test.ts by command family and promote mock context builders to testkit guard helpers.
+
+## Scope
+
+- In scope: Split guard commands.unit.test.ts by command family and promote mock context builders to testkit guard helpers.
+- Out of scope: unrelated refactors not required for "AP-09: Split guard implementation tests".
+
+## Verification
+
+### Plan
+
+1. Run `bun run test:project -- guard`. Expected: it succeeds and confirms the requested outcome for this task.
+2. Review the changed artifact or behavior for the `code` task. Expected: the requested outcome is visible and matches the approved scope.
+3. Compare the final result against the task summary and touched scope. Expected: remaining follow-up is either resolved or explicit in ## Findings.
+
+### Current Status
+
+- State: ok
+- Note: Verified: guard commands.unit.test.ts split into guard, commit close, and commit non-close suites; @agentplane/testkit/guard added for shared command context fixtures; guard project and oversized baseline pass.
+
+## Risks
+
+- Risk level: not recorded
+- Breaking change: no
+
+### Rollback
+
+- Revert task-related commit(s).
+- Re-run required checks to confirm rollback safety.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-01T09:29:51.147Z
+- Branch: task/202605010645-B5ERD0/guard-test-split
+- Head: 400d8fee5bd6
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
