+++ scripts/lib/test-route-registry.mjs
@@ -83,7 +83,9 @@ export const VITEST_WORKSPACE_PROJECTS = [
         "packages/agentplane/src/commands/guard/index.test.ts",
         "packages/agentplane/src/commands/guard/impl/allow.test.ts",
         "packages/agentplane/src/commands/guard/impl/comment-commit.test.ts",
-        "packages/agentplane/src/commands/guard/impl/commands.unit.test.ts",
+        "packages/agentplane/src/commands/guard/impl/commands.commit-close.unit.test.ts",
+        "packages/agentplane/src/commands/guard/impl/commands.commit-non-close.unit.test.ts",
+        "packages/agentplane/src/commands/guard/impl/commands.guard.unit.test.ts",
         "packages/agentplane/src/commands/guard/impl/policy.test.ts",
         "packages/agentplane/src/commands/guard/impl/close-message.test.ts",
       ],
@@ -311,7 +313,9 @@ const RELEASE_CRITICAL_FILES = [
 const SIGNIFICANT_COVERAGE_FILES = [
   "packages/agentplane/src/commands/guard/impl/allow.test.ts",
   "packages/agentplane/src/commands/guard/impl/close-message.test.ts",
-  "packages/agentplane/src/commands/guard/impl/commands.unit.test.ts",
+  "packages/agentplane/src/commands/guard/impl/commands.commit-close.unit.test.ts",
+  "packages/agentplane/src/commands/guard/impl/commands.commit-non-close.unit.test.ts",
+  "packages/agentplane/src/commands/guard/impl/commands.guard.unit.test.ts",
   "packages/agentplane/src/commands/guard/impl/policy.test.ts",
   "packages/agentplane/src/commands/guard/impl/comment-commit.test.ts",
   "packages/agentplane/src/cli/run-cli.core.guard.test.ts",
