+++ scripts/run-local-ci.mjs
@@ -154,7 +154,9 @@ const fullOnlySteps = [
         "run",
         "packages/agentplane/src/commands/guard/impl/allow.test.ts",
         "packages/agentplane/src/commands/guard/impl/close-message.test.ts",
-        "packages/agentplane/src/commands/guard/impl/commands.unit.test.ts",
+        "packages/agentplane/src/commands/guard/impl/commands.commit-close.unit.test.ts",
+        "packages/agentplane/src/commands/guard/impl/commands.commit-non-close.unit.test.ts",
+        "packages/agentplane/src/commands/guard/impl/commands.guard.unit.test.ts",
         "packages/agentplane/src/commands/guard/impl/policy.test.ts",
         "packages/agentplane/src/commands/guard/impl/comment-commit.test.ts",
         "packages/agentplane/src/cli/run-cli.core.guard.test.ts",
